pyproject.toml
[project]
name = "galaxtic"
version = "0.1.1"
description = "Discord bot for advanced server moderation and interactions."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py[voice]>=2.5.2",
    "langchain>=0.3.25",
    "pydantic-settings>=2.9.1",
    "rich>=14.0.0",
    "surrealdb>=1.0.4",
    "together>=1.5.8",
    "yt-dlp>=2025.5.22",
]
